move struct
