+++ src/Medulla/Kubernetes/Operator/Entities/V1Alpha1DataEntity.cs
@@ -10,42 +10,6 @@ using Operator.Model;
 namespace Medulla.Kubernetes.Operator.Entities;
 
 
-/// <summary>
-/// DatabaseSpec is spec of CRD for database
-/// </summary>
-public struct DatabaseSpec
-{
-    /// <summary>
-    /// Name is an element of DatabaseSpec which is the name of the database
-    /// </summary>
-    [Required]
-    public string? Name { get; set; }
-    /// <summary>
-    /// Host is an element of DatabaseSpec which is the connection path to the database
-    /// </summary>
-    [Required]
-    public string? Host { get; set; }
-    /// <summary>
-    /// Dialect is an element of DatabaseSpec which is the the dialect of the database
-    /// </summary>
-    [Required]
-    public string? Dialect { get; set; }
-    /// <summary>
-    /// UsernameSecretName is an element of DatabaseSpec which is the username to connect to the database
-    /// </summary>
-    [Required]
-    public string? UsernameSecretName { get; set; }
-    /// <summary>
-    /// PasswordSecretName is an element of DatabaseSpec which is the password to connect to the database
-    /// </summary>
-    [Required]
-    public string? PasswordSecretName { get; set; }
-    /// <summary>
-    /// Models is a list elements of type ModelSpec
-    /// </summary>
-    [Required]
-    public List<ModelSpec> Models { get; set; }
-}
